refactor(approvals): retire legacy channel_guardian_approval_requests subsystem (LUM-2497)
The legacy channel_guardian_approval_requests table lost its production writer
when channel reactions, callbacks, and text decisions migrated to the canonical
guardian pipeline, leaving every reader a no-op. Retire the whole dead subsystem:

- Delete the table CRUD (guardian-approvals.ts), the legacy callback strategy,
  the legacy expiry sweep, the legacy access-request decision path, and the
  now-empty channel-guardian-routes barrel.
- Remove the legacy applyGuardianDecision primitive; applyCanonicalGuardianDecision
  is the single guardian-decision write path.
- Collapse the dead branches in handleApprovalInterception (guardian callback
  dispatch, requester self-cancel, expiry auto-deny); keep the live unverified
  auto-deny and pre-row self-approval guards.
- Drop the table via migration 299.
- Migrate the notification-candidate pending-count query to the canonical table
  (it had silently returned empty since the canonical migration).
- Refresh stale docs (ARCHITECTURE.md, approvals/AGENTS.md) and tests; migrate
  the trusted-contact lifecycle-signal tests onto the canonical resolver.
