--- assistant/src/__tests__/access-request-decision.test.ts
@@ -1,375 +1,0 @@
-/**
- * Tests for the access request decision flow.
- *
- * When a guardian approves or denies an `ingress_access_request`:
- * - Approve: creates a verification session, delivers code to guardian,
- *   notifies requester to expect a code.
- * - Deny: sends refusal reply to requester.
- * - Stale: handles already-resolved requests gracefully.
- * - Idempotent: approving same request twice does not create duplicate sessions.
- */
-import { beforeEach, describe, expect, mock, test } from "bun:test";
-
-mock.module("../util/logger.js", () => ({
-  getLogger: () =>
-    new Proxy({} as Record<string, unknown>, {
-      get: () => () => {},
-    }),
-}));
-
-// Track deliverChannelReply calls and allow injecting failures
-const deliverReplyCalls: Array<{
-  url: string;
-  payload: Record<string, unknown>;
-}> = [];
-let deliverReplyError: Error | null = null;
-mock.module("../runtime/gateway-client.js", () => ({
-  deliverChannelReply: async (
-    url: string,
-    payload: Record<string, unknown>,
-  ) => {
-    if (deliverReplyError) {
-      throw deliverReplyError;
-    }
-    deliverReplyCalls.push({ url, payload });
-  },
-}));
-
-import { getDb } from "../memory/db-connection.js";
-import { initializeDb } from "../memory/db-init.js";
-import {
-  createApprovalRequest,
-  getApprovalRequestById,
-} from "../memory/guardian-approvals.js";
-import { findActiveSession } from "../runtime/channel-verification-service.js";
-import {
-  deliverVerificationCodeToGuardian,
-  handleAccessRequestDecision,
-  notifyRequesterOfApproval,
-  notifyRequesterOfDeliveryFailure,
-  notifyRequesterOfDenial,
-} from "../runtime/routes/access-request-decision.js";
-
-await initializeDb();
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-const GUARDIAN_APPROVAL_TTL_MS = 5 * 60 * 1000;
-
-function resetState(): void {
-  const db = getDb();
-  db.run("DELETE FROM channel_guardian_approval_requests");
-  db.run("DELETE FROM channel_verification_sessions");
-  deliverReplyCalls.length = 0;
-}
-
-function createTestApproval(overrides: Record<string, unknown> = {}) {
-  return createApprovalRequest({
-    runId: `ingress-access-request-${Date.now()}`,
-    conversationId: `access-req-telegram-user-unknown-456`,
-    channel: "telegram",
-    requesterExternalUserId: "user-unknown-456",
-    requesterChatId: "chat-123",
-    guardianExternalUserId: "guardian-user-789",
-    guardianChatId: "guardian-chat-789",
-    toolName: "ingress_access_request",
-    riskLevel: "access_request",
-    reason: "Alice Unknown is requesting access to the assistant",
-    expiresAt: Date.now() + GUARDIAN_APPROVAL_TTL_MS,
-    ...overrides,
-  });
-}
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-describe("access request decision handler", () => {
-  beforeEach(() => {
-    resetState();
-  });
-
-  test("guardian approve creates a verification session", () => {
-    const approval = createTestApproval();
-
-    const result = handleAccessRequestDecision(
-      approval,
-      "approve",
-      "guardian-user-789",
-    );
-
-    expect(result.handled).toBe(true);
-    expect(result.type).toBe("approved");
-    expect(result.verificationSessionId).toBeDefined();
-    expect(result.verificationCode).toBeDefined();
-    // Verification code should be a 6-digit numeric string
-    expect(result.verificationCode).toMatch(/^\d{6}$/);
-
-    // Approval record should be updated to 'approved'
-    const updated = getApprovalRequestById(approval.id);
-    expect(updated).not.toBeNull();
-    expect(updated!.status).toBe("approved");
-    expect(updated!.decidedByExternalUserId).toBe("guardian-user-789");
-  });
-
-  test("verification session is identity-bound to the requester", () => {
-    const approval = createTestApproval();
-
-    const result = handleAccessRequestDecision(
-      approval,
-      "approve",
-      "guardian-user-789",
-    );
-
-    expect(result.type).toBe("approved");
-
-    // There should be an active session for this channel
-    const session = findActiveSession("telegram");
-    expect(session).not.toBeNull();
-    expect(session!.expectedExternalUserId).toBe("user-unknown-456");
-    expect(session!.expectedChatId).toBe("chat-123");
-    expect(session!.identityBindingStatus).toBe("bound");
-    expect(session!.status).toBe("awaiting_response");
-  });
-
-  test("guardian deny marks approval as denied", () => {
-    const approval = createTestApproval();
-
-    const result = handleAccessRequestDecision(
-      approval,
-      "deny",
-      "guardian-user-789",
-    );
-
-    expect(result.handled).toBe(true);
-    expect(result.type).toBe("denied");
-    expect(result.verificationSessionId).toBeUndefined();
-    expect(result.verificationCode).toBeUndefined();
-
-    // Approval record should be updated to 'denied'
-    const updated = getApprovalRequestById(approval.id);
-    expect(updated).not.toBeNull();
-    expect(updated!.status).toBe("denied");
-    expect(updated!.decidedByExternalUserId).toBe("guardian-user-789");
-
-    // No verification session should be created
-    const session = findActiveSession("telegram");
-    expect(session).toBeNull();
-  });
-
-  test("stale decision (already resolved) returns stale", () => {
-    const approval = createTestApproval();
-
-    // Approve first
-    handleAccessRequestDecision(approval, "approve", "guardian-user-789");
-
-    // Try to deny the same approval — should be stale
-    const result = handleAccessRequestDecision(
-      approval,
-      "deny",
-      "guardian-user-789",
-    );
-
-    expect(result.handled).toBe(true);
-    expect(result.type).toBe("stale");
-  });
-
-  test("idempotent approval does not create duplicate verification sessions", () => {
-    const approval = createTestApproval();
-
-    // Approve first
-    const result1 = handleAccessRequestDecision(
-      approval,
-      "approve",
-      "guardian-user-789",
-    );
-    expect(result1.type).toBe("approved");
-    const _sessionId1 = result1.verificationSessionId;
-
-    // Approve again — should be idempotent (already resolved with same decision)
-    const result2 = handleAccessRequestDecision(
-      approval,
-      "approve",
-      "guardian-user-789",
-    );
-
-    // resolveApprovalRequest returns the existing record for same-decision idempotency,
-    // but since the approval is no longer 'pending', a second createOutboundSession
-    // will still be called. However, createOutboundSession auto-revokes prior sessions,
-    // so there will be exactly one active session at the end.
-    // The important thing is that the result indicates approval was handled.
-    expect(result2.handled).toBe(true);
-    // Either 'approved' (creates a new session) or something else is acceptable,
-    // but it should not crash.
-  });
-});
-
-describe("access request notification delivery", () => {
-  beforeEach(() => {
-    deliverReplyCalls.length = 0;
-    deliverReplyError = null;
-  });
-
-  test("delivers verification code to guardian and returns ok", async () => {
-    const result = await deliverVerificationCodeToGuardian({
-      replyCallbackUrl: "http://localhost:7830/deliver/telegram",
-      guardianChatId: "guardian-chat-789",
-      requesterIdentifier: "user-unknown-456",
-      verificationCode: "123456",
-      assistantId: "self",
-    });
-
-    expect(result.ok).toBe(true);
-    expect(deliverReplyCalls.length).toBe(1);
-    const call = deliverReplyCalls[0];
-    expect(call.payload.chatId).toBe("guardian-chat-789");
-    const text = call.payload.text as string;
-    expect(text).toContain("123456");
-    expect(text).toContain("user-unknown-456");
-    expect(text).toContain("10 minutes");
-  });
-
-  test("returns failure result when guardian code delivery fails", async () => {
-    deliverReplyError = new Error("Gateway timeout");
-
-    const result = await deliverVerificationCodeToGuardian({
-      replyCallbackUrl: "http://localhost:7830/deliver/telegram",
-      guardianChatId: "guardian-chat-789",
-      requesterIdentifier: "user-unknown-456",
-      verificationCode: "123456",
-      assistantId: "self",
-    });
-
-    expect(result.ok).toBe(false);
-    if (!result.ok) {
-      expect(result.reason).toBe("Gateway timeout");
-    }
-    // No calls should have been recorded (error thrown before push)
-    expect(deliverReplyCalls.length).toBe(0);
-  });
-
-  test("notifies requester of approval", async () => {
-    await notifyRequesterOfApproval({
-      replyCallbackUrl: "http://localhost:7830/deliver/telegram",
-      requesterChatId: "chat-123",
-      assistantId: "self",
-    });
-
-    expect(deliverReplyCalls.length).toBe(1);
-    const call = deliverReplyCalls[0];
-    expect(call.payload.chatId).toBe("chat-123");
-    const text = call.payload.text as string;
-    expect(text).toContain("approved");
-    expect(text).toContain("verification code");
-  });
-
-  test("notifies requester of denial", async () => {
-    await notifyRequesterOfDenial({
-      replyCallbackUrl: "http://localhost:7830/deliver/telegram",
-      requesterChatId: "chat-123",
-      assistantId: "self",
-    });
-
-    expect(deliverReplyCalls.length).toBe(1);
-    const call = deliverReplyCalls[0];
-    expect(call.payload.chatId).toBe("chat-123");
-    const text = call.payload.text as string;
-    expect(text).toContain("denied");
-  });
-
-  test("notifies requester of delivery failure", async () => {
-    await notifyRequesterOfDeliveryFailure({
-      replyCallbackUrl: "http://localhost:7830/deliver/telegram",
-      requesterChatId: "chat-123",
-      assistantId: "self",
-    });
-
-    expect(deliverReplyCalls.length).toBe(1);
-    const call = deliverReplyCalls[0];
-    expect(call.payload.chatId).toBe("chat-123");
-    const text = call.payload.text as string;
-    expect(text).toContain("approved");
-    expect(text).toContain("unable to deliver");
-    expect(text).toContain("try again");
-  });
-
-  test("slack approval notification is sent as DM using requesterExternalUserId", async () => {
-    await notifyRequesterOfApproval({
-      replyCallbackUrl:
-        "http://localhost:7830/deliver/slack?threadTs=1234.5678",
-      requesterChatId: "C12345-channel",
-      requesterExternalUserId: "U98765-user",
-      channel: "slack",
-      assistantId: "self",
-    });
-
-    expect(deliverReplyCalls.length).toBe(1);
-    const call = deliverReplyCalls[0];
-    // Should target the user ID (DM) not the channel
-    expect(call.payload.chatId).toBe("U98765-user");
-    // threadTs should be stripped — it belongs to the guardian's channel thread
-    expect(call.url).not.toContain("threadTs");
-  });
-
-  test("slack denial notification is sent as DM using requesterExternalUserId", async () => {
-    await notifyRequesterOfDenial({
-      replyCallbackUrl:
-        "http://localhost:7830/deliver/slack?threadTs=1234.5678",
-      requesterChatId: "C12345-channel",
-      requesterExternalUserId: "U98765-user",
-      channel: "slack",
-      assistantId: "self",
-    });
-
-    expect(deliverReplyCalls.length).toBe(1);
-    const call = deliverReplyCalls[0];
-    expect(call.payload.chatId).toBe("U98765-user");
-    expect(call.url).not.toContain("threadTs");
-  });
-
-  test("slack delivery failure notification is sent as DM using requesterExternalUserId", async () => {
-    await notifyRequesterOfDeliveryFailure({
-      replyCallbackUrl:
-        "http://localhost:7830/deliver/slack?threadTs=1234.5678",
-      requesterChatId: "C12345-channel",
-      requesterExternalUserId: "U98765-user",
-      channel: "slack",
-      assistantId: "self",
-    });
-
-    expect(deliverReplyCalls.length).toBe(1);
-    const call = deliverReplyCalls[0];
-    expect(call.payload.chatId).toBe("U98765-user");
-    expect(call.url).not.toContain("threadTs");
-  });
-
-  test("non-slack channels still use requesterChatId and preserve threadTs", async () => {
-    await notifyRequesterOfApproval({
-      replyCallbackUrl:
-        "http://localhost:7830/deliver/telegram?threadTs=1234.5678",
-      requesterChatId: "chat-123",
-      requesterExternalUserId: "user-456",
-      channel: "telegram",
-      assistantId: "self",
-    });
-
-    expect(deliverReplyCalls.length).toBe(1);
-    expect(deliverReplyCalls[0].payload.chatId).toBe("chat-123");
-    // threadTs should be preserved for non-slack channels
-    expect(deliverReplyCalls[0].url).toContain("threadTs=1234.5678");
-  });
-
-  test("slack without requesterExternalUserId falls back to requesterChatId", async () => {
-    await notifyRequesterOfApproval({
-      replyCallbackUrl: "http://localhost:7830/deliver/slack",
-      requesterChatId: "C12345-channel",
-      channel: "slack",
-      assistantId: "self",
-    });
-
-    expect(deliverReplyCalls.length).toBe(1);
-    expect(deliverReplyCalls[0].payload.chatId).toBe("C12345-channel");
-  });
-});
